requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
torch
